+++ src/main/java/com/github/nemiro54/telegrambot/TelegramBotApplication.java
@@ -2,8 +2,10 @@ package com.github.nemiro54.telegrambot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class TelegramBotApplication {
 
 	public static void main(String[] args) {