+++ src/main/java/com/github/nemiro54/telegrambot/javarushclient/dto/Language.java
@@ -1,0 +1,31 @@
+package com.github.nemiro54.telegrambot.javarushclient.dto;
+
+/**
+ * DTO, which represents languages.
+ */
+public enum Language {
+
+    UNKNOWN,
+    ENGLISH,
+    GERMAN,
+    SPANISH,
+    HINDI,
+    FRENCH,
+    PORTUGUESE,
+    POLISH,
+    BENGALI,
+    PUNJABI,
+    CHINESE,
+    ITALIAN,
+    INDONESIAN,
+    MARATHI,
+    TAMIL,
+    TELUGU,
+    JAPANESE,
+    KOREAN,
+    URDU,
+    TAIWANESE,
+    NETHERLANDS,
+    RUSSIAN,
+    UKRAINIAN
+}