+++ src/main/java/com/github/nemiro54/telegrambot/javarushclient/dto/PostType.java
@@ -1,0 +1,12 @@
+package com.github.nemiro54.telegrambot.javarushclient.dto;
+
+/**
+ * DTO, which represents post types.
+ */
+public enum PostType {
+
+    UNKNOWN,
+    USUAL,
+    INNER_LINK,
+    OUTER_LINK
+}