+++ src/main/java/com/github/nemiro54/telegrambot/javarushclient/dto/VisibilityStatus.java
@@ -1,0 +1,15 @@
+package com.github.nemiro54.telegrambot.javarushclient.dto;
+
+/**
+ * DTO, which represents visibility status.
+ */
+public enum VisibilityStatus {
+
+    UNKNOWN,
+    RESTRICTED,
+    PUBLIC,
+    PROTECTED,
+    PRIVATE,
+    DISABLED,
+    DELETED
+}