+++ src/main/java/com/github/nemiro54/telegrambot/service/SendBotMessageService.java
@@ -1,5 +1,7 @@
 package com.github.nemiro54.telegrambot.service;
 
+import java.util.List;
+
 /**
  * Service for sending messages via telegram bot.
  */
@@ -8,8 +10,16 @@ public interface SendBotMessageService {
     /**
      * Send message via telegram bot.
      *
-     * @param chatId provided chatId in which messages would be sent.
+     * @param chatId provided chatId in which would be sent.
      * @param message provided message to be sent.
      */
     void sendMessage(String chatId, String message);
+
+    /**
+     * Send message via telegram bot.
+     *
+     * @param chatId provided chatId in which would be sent.
+     * @param message collection of provided messages to be sent.
+     */
+    void sendMessage(String chatId, List<String> message);
 }